+++ src/tydier/__init__.py
@@ -1,5 +1,10 @@
+# This module is part of the `tydier` project. Please find more information
+# at https://github.com/antobzzll/tydier
+
 # read version from installed package
 from importlib.metadata import version
 __version__ = version("tydier")
+
+
 
 # from .tydier import *