+++ src/tydier/catvars.py
@@ -1,3 +1,6 @@
+# This module is part of the `tydier` project. Please find more information
+# at https://github.com/antobzzll/tydier
+
 import pandas as pd
 from .strings import *
 import warnings