+++ src/tydier/__init__.py
@@ -6,5 +6,6 @@ from importlib.metadata import version
 __version__ = version("tydier")
 
 
-
-# from .tydier import *
+from .catvars import *
+from .numvars import *
+from .strings import *