+++ src/tydier/numvars.py
@@ -1,8 +1,8 @@
 # This module is part of the `tydier` project. Please find more information
 # at https://github.com/antobzzll/tydier
 
-import pandas as pd
 from .strings import *
+import pandas as pd
 import warnings
 warnings.simplefilter(action='ignore', category=FutureWarning)